tests/wpt/web-platform-tests/dom/nodes/Document-contentType/contentType/contenttype_png.html

<!DOCTYPE html>
<title>PNG document.contentType === 'image/png'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe style="display:none"></iframe>
<script>
async_test(function() {
  var iframe = document.getElementsByTagName('iframe')[0];
  iframe.addEventListener('load', this.step_func_done(function() {
    assert_equals(iframe.contentDocument.contentType, "image/png");
  }), false);
  iframe.src = "../resources/t.png";
});
</script>
